part-6

+++ src/tasks/tasks.service.ts
@@ -14,12 +14,14 @@ export class TasksService {
     private taskRepository: TaskRepository
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto) {
-    return await this.taskRepository.getTasks(filterDto)
+  async getTasks(filterDto: GetTasksFilterDto, user: User) {
+    return await this.taskRepository.getTasks(filterDto, user)
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id)
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    })
 
     if (!found) {
       throw new NotFoundException(`This task with id: ${id} not found!`)
@@ -32,39 +34,22 @@ export class TasksService {
     return this.taskRepository.createTask(createTaskDto, user)
   }
 
-  async deleteById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id)
+  async deleteById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id })
 
     if (result.affected === 0) {
       throw new NotFoundException(`This task with id: ${id} not found!`)
     }
   }
 
-  async patchStatusById(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id)
+  async patchStatusById(
+    id: number,
+    status: TaskStatus,
+    user: User
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     task.status = status
     await task.save()
     return task
   }
-
-  // getAllTasks(): Task[] {
-  //     return this.tasks
-  // }
-
-  // getTasksWithFilter( filterDto: GetTasksFilterDto) {
-  //     const { status, search } = filterDto
-  //     let tasks = this.getAllTasks()
-
-  //     if ( status ) {
-  //         tasks = tasks.filter( item => item.status === status)
-  //     }
-
-  //     if ( search ) {
-  //         tasks = tasks.filter( item =>
-  //             item.title.includes( search )
-  //             || item.description.includes( search ))
-  //     }
-
-  //     return tasks
-  // }
 }
